FastAPI/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-openai
langchain-community
youtube-transcript-api
python-dotenv
faiss-cpu
tiktoken
openai